+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Implemented offset based storage access (#843)
 
+## 0.5.1 (2024-08-28) - `miden-objects` crate only
+
+- Implemented `PrettyPrint` and `Display` for `NoteScript`.
+
 ## 0.5.0 (2024-08-27)
 
 ### Features